kimi-security-auditor/pyproject.toml
[project]
name = "kimi-security-auditor"
version = "0.1.0"
description = "A comprehensive web security auditing tool"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Kimi Security Team"}
]
keywords = ["security", "auditing", "penetration-testing", "web-security", "scanner"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
kimi-audit = "kimi_security_auditor.cli:main"

[project.urls]
Homepage = "https://github.com/kimi-ecosystem/kimi-security-auditor"
Repository = "https://github.com/kimi-ecosystem/kimi-security-auditor"
Issues = "https://github.com/kimi-ecosystem/kimi-security-auditor/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kimi_security_auditor"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
